+++ src/app/app.module.ts
@@ -10,7 +10,6 @@ import { SharedModule } from './shared/shared.module';
 @NgModule({
   declarations: [AppComponent],
   imports: [BrowserModule, CartModule, ProductsModule, OrdersModule, SharedModule],
-  providers: [],
   bootstrap: [AppComponent],
 })
 export class AppModule {}