+++ src/app/cart/services/cart.service.ts
@@ -5,24 +5,25 @@ import { ICartItemData } from '../models/cart.models';
   providedIn: 'root',
 })
 export class CartService {
-  cartList: ICartItemData[] = [];
-  totalAmount = 0;
-  numberOfItems = 0;
+  cartProducts: ICartItemData[] = [];
+  totalSum = 0;
+  totalQuantity = 0;
+  isEmptyCart = true;
 
-  getCartItems(): ICartItemData[] {
-    return this.cartList;
+  getProducts(): ICartItemData[] {
+    return this.cartProducts;
   }
 
-  addItemToCartList(id: number, name: string, price: number): void {
-    const itemToAdd = this.cartList.find((item) => item.id === id);
+  addProduct(id: number, name: string, price: number, quantity = 1): void {
+    const itemToAdd = this.cartProducts.find((item) => item.id === id);
 
     if (!itemToAdd) {
-      this.cartList = [
-        ...this.cartList,
-        { id: id, name: name, quantity: 1, price: price, amount: price },
+      this.cartProducts = [
+        ...this.cartProducts,
+        { id: id, name: name, quantity: quantity, price: price, amount: price },
       ];
     } else {
-      this.cartList = this.cartList.map((item) => {
+      this.cartProducts = this.cartProducts.map((item) => {
         const itemCopy = { ...item };
         if (itemCopy.id !== id) {
           return itemCopy;
@@ -33,75 +34,71 @@ export class CartService {
         }
       });
     }
-
-    this.calculateCartAmount();
-    this.calculateNumberOfItems();
+    this.updateCartData();
   }
 
   getItemAmount(quantity: number, price: number): number {
     return Number((quantity * price).toFixed(2));
   }
 
-  private calculateCartAmount(): void {
-    this.totalAmount = Number(
-      this.cartList
+  private updateCartData(): void {
+    this.totalSum = Number(
+      this.cartProducts
         .map((item) => item.amount)
         .reduce((a, b) => a + b, 0)
         .toFixed(2)
     );
-  }
-
-  private calculateNumberOfItems(): void {
-    this.numberOfItems = this.cartList.map((item) => item.quantity).reduce((a, b) => a + b, 0);
+    this.totalQuantity = this.cartProducts.map((item) => item.quantity).reduce((a, b) => a + b, 0);
+    this.isEmptyCart = this.totalQuantity > 0;
   }
 
   getTotalCartAmount(): number {
-    return this.totalAmount;
+    return this.totalSum;
   }
 
   getNumberOfItems(): number {
-    return this.numberOfItems;
+    return this.totalQuantity;
   }
 
-  increaseItemQuantity(id: number): void {
-    this.cartList = this.cartList.map((item) => {
+  increaseQuantity(id: number): void {
+    this.cartProducts = this.cartProducts.map((item) => {
       let itemCopy = item;
 
       if (itemCopy.id === id) {
-        itemCopy.quantity += 1;
-        itemCopy.amount = this.getItemAmount(itemCopy.quantity, itemCopy.price);
+        itemCopy = this.changeQuantity(itemCopy, 1);
       }
       return itemCopy;
     });
-
-    this.calculateCartAmount();
-    this.calculateNumberOfItems();
+    this.updateCartData();
   }
 
-  decreaseItemQuantity(id: number): void {
-    this.cartList = this.cartList.map((item) => {
+  decreaseQuantity(id: number): void {
+    this.cartProducts = this.cartProducts.map((item) => {
       let itemCopy = item;
 
       if (itemCopy.id === id) {
         if (itemCopy.quantity > 1) {
-          itemCopy.quantity -= 1;
-          itemCopy.amount = this.getItemAmount(itemCopy.quantity, itemCopy.price);
+          itemCopy = this.changeQuantity(itemCopy, -1);
         }
       }
       return itemCopy;
     });
-
-    this.calculateCartAmount();
-    this.calculateNumberOfItems();
+    this.updateCartData();
   }
 
-  deleteItemFromCartList(id: number): void {
-    this.cartList.splice(
-      this.cartList.findIndex((item) => item.id === id),
-      1
-    );
+  private changeQuantity(item: ICartItemData, quantityDelta: number): ICartItemData {
+    const itemCopy = { ...item };
+    itemCopy.quantity += quantityDelta;
+    itemCopy.amount = this.getItemAmount(itemCopy.quantity, itemCopy.price);
+    return itemCopy;
+  }
 
-    this.calculateCartAmount();
-    this.calculateNumberOfItems();
+  removeProduct(id: number): void {
+    this.cartProducts = this.cartProducts.filter((item) => item.id !== id);
+    this.updateCartData();
+  }
+
+  removeAllProducts(): void {
+    this.cartProducts = [];
   }
 }