+++ src/app/core/services/generator/generator.factory.ts
@@ -1,0 +1,5 @@
+import { GeneratorService } from './generator.service';
+
+export function GeneratorFactory(n = 6): (generatorService: GeneratorService) => string {
+  return (generatorService: GeneratorService): string => generatorService.generate(n);
+}